+++ dbms/src/Common/Config/CMakeLists.txt
@@ -1,0 +1,9 @@
+
+include(${ClickHouse_SOURCE_DIR}/cmake/dbms_glob_sources.cmake)
+
+add_headers_and_sources(clickhouse_common_config .)
+
+add_library(clickhouse_common_config ${SPLIT_SHARED} ${clickhouse_common_config_headers} ${clickhouse_common_config_sources})
+
+target_link_libraries (clickhouse_common_config clickhouse_common_zookeeper string_utils)
+target_include_directories (clickhouse_common_config PRIVATE ${DBMS_INCLUDE_DIR})
