+++ dbms/src/Common/Config/ConfigReloader.cpp
@@ -2,11 +2,9 @@
 
 #include <Poco/Util/Application.h>
 #include <Poco/File.h>
-
 #include <common/logger_useful.h>
-
 #include <Common/setThreadName.h>
-#include <Common/ConfigProcessor/ConfigProcessor.h>
+#include "ConfigProcessor.h"
 
 
 namespace DB
