+++ dbms/src/Common/ZooKeeper/tests/CMakeLists.txt
@@ -14,4 +14,4 @@ add_executable(zkutil_zookeeper_holder zkutil_zookeeper_holder.cpp)
 target_link_libraries(zkutil_zookeeper_holder clickhouse_common_zookeeper)
 
 add_executable (zk_many_watches_reconnect zk_many_watches_reconnect.cpp)
-target_link_libraries (zk_many_watches_reconnect clickhouse_common_zookeeper clickhouse_common_configprocessor)
+target_link_libraries (zk_many_watches_reconnect clickhouse_common_zookeeper clickhouse_common_config)
