+++ libs/libdaemon/include/daemon/BaseDaemon.h
@@ -23,7 +23,7 @@
 #include <common/Types.h>
 #include <common/logger_useful.h>
 #include <daemon/GraphiteWriter.h>
-#include <Common/ConfigProcessor/ConfigProcessor.h>
+#include <Common/Config/ConfigProcessor.h>
 
 namespace Poco { class TaskManager; }
